+++ packages/valory/skills/market_creation_manager_abci/behaviours/deposit_dai.py
@@ -28,7 +28,7 @@ from packages.valory.skills.market_creation_manager_abci.behaviours.base import 
     MarketCreationManagerBaseBehaviour,
     ETHER_VALUE
 )
-from packages.valory.skills.market_creation_manager_abci.rounds import DepositDaiRound
+from packages.valory.skills.market_creation_manager_abci.states.deposit_dai_round import DepositDaiRound
 from packages.valory.skills.market_creation_manager_abci.payloads import DepositDaiPayload
 from packages.valory.skills.transaction_settlement_abci.payload_tools import hash_payload_to_hex
